removes BaseFilemanagerWidget, fixes set title and handlerUrl on widget init config
