+++ src/widgets/FileManagerWidget.php
@@ -9,17 +9,19 @@
 
 namespace hrzg\filemanager\widgets;
 
-use hrzg\filemanager\widgets\base\BaseFileManagerWidget;
-
+use hrzg\filemanager\assets\AfmAsset;
+use yii\base\Widget;
+use yii\helpers\Url;
+use yii\web\View;
 
 /**
- * CLass FileManagerWidget
- * File manager widget
+ * Class FileManagerWidget
+ * File manager base widget
  *
- * @package hrzg\filemanager\widgets
+ * @package hrzg\filemanager\widgets\base
  * @author Christopher Stebe <c.stebe@herzogkommunikation.de>
  */
-class FileManagerWidget extends BaseFileManagerWidget
+class FileManagerWidget extends Widget
 {
     /**
      * File Handler Url
@@ -30,7 +32,7 @@ class FileManagerWidget extends BaseFileManagerWidget
     /**
      * @var string
      */
-    public $title = 'My Filemanager';
+    public $title = '';
 
     /**
      * @var string
@@ -38,10 +40,110 @@ class FileManagerWidget extends BaseFileManagerWidget
     public $template = "<div data-ng-app=\"FileManagerApp\"><div class=\"ng-cloak\"><angular-filemanager></angular-filemanager></div></div>";
 
     /**
+     * @inheritdoc
+     */
+    public function init()
+    {
+        // register assets
+        AfmAsset::register(\Yii::$app->view);
+
+        // Config filemanager
+        $this->setFileManagerConfig();
+    }
+
+    /**
      * @return string
      */
     public function run()
     {
         return $this->template;
+    }
+
+    /**
+     * Set handler url and init angular module config
+     */
+    protected function setFileManagerConfig()
+    {
+        // Set handler Url
+        if ($this->handlerUrl === null) {
+            $this->handlerUrl = Url::to([getenv('AFM_HANDLER_URL')]);
+        }
+
+        // check if user is allowed to change permissions
+        $allowPermissions = false;
+        if (!\Yii::$app->user->isGuest && \Yii::$app->user->can('FileflyPermissions')) {
+            $allowPermissions = true;
+        }
+
+        $title             = empty($this->title) ? getenv('AFM_TITLE') : $this->title;
+        $lang              = \Yii::$app->language;
+        $initFilemanagerJs = <<<JS
+angular.module('FileManagerApp').config(['fileManagerConfigProvider', function (config) {
+    var defaults = config.\$get();
+    var handler = '$this->handlerUrl';
+    config.set({
+
+        // Application
+        appName: '$title',
+        defaultLang: '$lang',
+        searchForm: true,
+        sidebar: true,
+        breadcrumb: true,
+        hidePermissions: true,
+
+        // Allowed actions
+        allowedActions: angular.extend(defaults.allowedActions, {
+            remove: true,
+            list: true,
+            move: true,
+            rename: true,
+            copy: true,
+            download: true,
+            downloadMultiple: false,
+            downloadLink: true,
+            changePermissions: '$allowPermissions',
+            compress: false,
+            compressChooseName: false,
+            extract: false,
+            upload: true
+        }),
+
+        // Handler
+        listUrl: handler,
+        uploadUrl: handler,
+        renameUrl: handler,
+        copyUrl: handler,
+        moveUrl: handler,
+        removeUrl: handler,
+        getContentUrl: handler,
+        createFolderUrl: handler,
+        downloadFileUrl: handler,
+        downloadMultipleUrl: handler,
+        compressUrl: handler,
+        extractUrl: handler,
+        permissionsUrl: handler,
+
+        // Additional settings
+        multipleDownloadFileName: 'filemanager.zip',
+        showSizeForDirectories: false,
+        useBinarySizePrefixes: false,
+        downloadFilesByAjax: true,
+        previewImagesInModal: true,
+        enablePermissionsRecursive: false,
+
+        // File patterns
+        isEditableFilePattern: /\.(!)/i,
+        isImageFilePattern: /\.(jpe?g|gif|bmp|png|svg|tiff?)$/i,
+        isExtractableFilePattern: /\.(gz|tar|rar|g?zip)$/i
+        //tplPath: 'src/templates'
+    });
+}]);
+JS;
+
+        // Register
+        \Yii::$app->view->registerJs(
+            $initFilemanagerJs,
+            View::POS_HEAD
+        );
     }
 }
